+++ CHANGELOG.md
@@ -9,7 +9,7 @@ The following is a curated list of changes in the Enact agate module, newest cha
 - `agate/Heading` prop `showBackButton`
 - `agate/Input` prop `clearButton` and `clearIcon`
 - `agate/Keypad` prop `activeCall` to render different icons depending on whether there is an active call or not
-- `agate/MediaPlayer` prop `type` to have a `mini` mediaPlayer
+- `agate/MediaPlayer` prop `type`
 
 ### Fixed